Futzing to get some reliability on OSX
On OSX I get frequent issues with DNS which stop the automation of the install (eg, fail to download kickstart, fail to download install.img).  Unfortunately using NAT appears to be the only really generic way to get to the VM (via port forwarding) to allow veewee to work, so this solution is likely very environment specific.

The Veeweefile is purely to undo the extra interface I'm using to allow reliable networking between the guest and host.